Unify journal commands: checkpoint and quorum
