+++ shell/src/main/java/alluxio/cli/fsadmin/journal/QuorumInfoCommand.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.cli.fsadmin.quorum;
+package alluxio.cli.fsadmin.journal;
 
 import alluxio.cli.fsadmin.command.AbstractFsAdminCommand;
 import alluxio.cli.fsadmin.command.Context;
@@ -17,8 +17,9 @@ import alluxio.client.journal.JournalMasterClient;
 import alluxio.conf.AlluxioConfiguration;
 import alluxio.exception.ExceptionMessage;
 import alluxio.exception.status.InvalidArgumentException;
+import alluxio.grpc.GetQuorumInfoPResponse;
 import alluxio.grpc.JournalDomain;
-import alluxio.grpc.NetAddress;
+import alluxio.grpc.QuorumServerInfo;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.commons.cli.CommandLine;
@@ -28,20 +29,20 @@ import java.io.IOException;
 import java.util.Arrays;
 
 /**
- * Command for removing a server from journal quorum.
+ * Command for querying journal quorum information.
  */
-public class QuorumRemoveCommand extends AbstractFsAdminCommand {
-
-  public static final String ADDRESS_OPTION_NAME = "address";
+public class QuorumInfoCommand extends AbstractFsAdminCommand {
   public static final String DOMAIN_OPTION_NAME = "domain";
 
-  public static final String OUTPUT_RESULT = "Removed server at: %s from quorum: %s";
+  public static final String OUTPUT_HEADER_DOMAIN = "Journal domain\t: %s";
+  public static final String OUTPUT_HEADER_QUORUM_SIZE = "Quorum size\t: %d";
+  public static final String OUTPUT_SERVER_INFO = "%s\t %s:%s";
 
   /**
    * @param context fsadmin command context
    * @param alluxioConf Alluxio configuration
    */
-  public QuorumRemoveCommand(Context context, AlluxioConfiguration alluxioConf) {
+  public QuorumInfoCommand(Context context, AlluxioConfiguration alluxioConf) {
     super(context);
   }
 
@@ -50,7 +51,7 @@ public class QuorumRemoveCommand extends AbstractFsAdminCommand {
    */
   @VisibleForTesting
   public static String description() {
-    return "Removes a server from embedded journal quorum.";
+    return "Shows quorum information for embedded journal.";
   }
 
   @Override
@@ -64,35 +65,32 @@ public class QuorumRemoveCommand extends AbstractFsAdminCommand {
       }
     } catch (IllegalArgumentException e) {
       throw new InvalidArgumentException(ExceptionMessage.INVALID_OPTION_VALUE
-              .getMessage(DOMAIN_OPTION_NAME, Arrays.toString(JournalDomain.values())));
+          .getMessage(DOMAIN_OPTION_NAME, Arrays.toString(JournalDomain.values())));
     }
 
-    String serverAddress = cl.getOptionValue(ADDRESS_OPTION_NAME);
-    // Extract hostname:port from given address.
-    String hostName;
-    int port;
-    try {
-      hostName = serverAddress.substring(0, serverAddress.indexOf(":"));
-      port = Integer.parseInt(serverAddress.substring(serverAddress.indexOf(":") + 1));
-    } catch (Exception e) {
-      throw new InvalidArgumentException(ExceptionMessage.INVALID_ADDRESS_VALUE.getMessage());
+    GetQuorumInfoPResponse quorumInfo = jmClient.getQuorumInfo();
+    mPrintStream.println(String.format(OUTPUT_HEADER_DOMAIN, quorumInfo.getDomain()));
+    mPrintStream
+        .println(String.format(OUTPUT_HEADER_QUORUM_SIZE, quorumInfo.getServerInfoList().size()));
+    mPrintStream.println("STATE\t\tSERVER ADDRESS");
+    for (QuorumServerInfo serverState : quorumInfo.getServerInfoList()) {
+      String serverStateStr = String.format(OUTPUT_SERVER_INFO, serverState.getServerState(),
+          serverState.getServerAddress().getHost(), serverState.getServerAddress().getRpcPort());
+      mPrintStream.println(serverStateStr);
     }
-
-    jmClient.removeQuorumServer(NetAddress.newBuilder().setHost(hostName).setRpcPort(port).build());
-    mPrintStream.println(String.format(OUTPUT_RESULT, serverAddress, domainVal));
 
     return 0;
   }
 
   @Override
   public String getCommandName() {
-    return "remove";
+    return "info";
   }
 
   @Override
   public String getUsage() {
-    return String.format("%s -%s <MASTER|JOB_MASTER>%n -%s <HostName:Port>", getCommandName(),
-        DOMAIN_OPTION_NAME, ADDRESS_OPTION_NAME);
+    return String.format("%s -%s <%s>%n", getCommandName(), DOMAIN_OPTION_NAME,
+        Arrays.toString(JournalDomain.values()));
   }
 
   @Override
@@ -102,20 +100,19 @@ public class QuorumRemoveCommand extends AbstractFsAdminCommand {
 
   @Override
   public void validateArgs(CommandLine cl) throws InvalidArgumentException {
-    if (cl.getOptions().length != 2) {
+    if (cl.getOptions().length != 1) {
       throw new InvalidArgumentException(
-          ExceptionMessage.INVALID_OPTION_COUNT.getMessage(2, cl.getOptions().length));
+          ExceptionMessage.INVALID_OPTION_COUNT.getMessage(1, cl.getOptions().length));
     }
     // Validate passed options are correct.
-    if (!cl.hasOption(DOMAIN_OPTION_NAME) || !cl.hasOption(ADDRESS_OPTION_NAME)) {
-      throw new InvalidArgumentException(ExceptionMessage.INVALID_OPTION
-          .getMessage(String.format("[%s, %s]", DOMAIN_OPTION_NAME, ADDRESS_OPTION_NAME)));
+    if (!cl.hasOption(DOMAIN_OPTION_NAME)) {
+      throw new InvalidArgumentException(
+          ExceptionMessage.INVALID_OPTION.getMessage(String.format("[%s]", DOMAIN_OPTION_NAME)));
     }
   }
 
   @Override
   public Options getOptions() {
-    return new Options().addOption(DOMAIN_OPTION_NAME, true, "Journal domain")
-        .addOption(ADDRESS_OPTION_NAME, true, "Server address to remove");
+    return new Options().addOption(DOMAIN_OPTION_NAME, true, "Journal domain");
   }
 }
